+++ otreeutils/__init__.py
@@ -1,4 +1,5 @@
 __title__ = 'otreeutils'
-__version__ = '0.3.1'
+__version__ = '0.4.0dev'
 __author__ = 'Markus Konrad'
 __license__ = 'Apache License 2.0'
+