+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## v0.4.0 – 2018-09-27
+
+* added admin extensions:
+    * live session data with custom data models
+    * app data export with custom data models in CSV, Excel and JSON formats
+* dropped support for oTree v1.x
+* fixed some minor compat. issues with latest oTree version
+
 ## v0.3.0 – 2018-04-25
 
 * made compatible with oTree v2.0